pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stream"
version = "1.0.0"
description = "STREAM - Smart Tiered Routing Engine for AI Models"
requires-python = ">=3.11"
authors = [
    {name = "Anas Nassar", email = "nassar@uic.edu"}
]

dependencies = [
    # Core Python utilities
    "python-dotenv>=1.0.0",
    # AI & LLM
    "litellm>=1.80.11",
    "openai>=2.14.0",
    # UI (Streamlit)
    "streamlit>=1.52.2",
    # Middleware - Web Framework
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.4",
    # Middleware - HTTP & Networking
    "httpx>=0.28.1",
    "python-multipart>=0.0.20",
    # Database
    "psycopg2-binary>=2.9.10",
    # Utilities
    "rich>=14.2.0",
    "tiktoken>=0.12.0",
    "python-json-logger>=4.0.0",
    "pyyaml>=6.0.3",
    "globus-compute-sdk>=4.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "ruff>=0.8.0",
]

# Tell hatchling where the package is
[tool.hatch.build.targets.wheel]
packages = ["stream"]

# Entry points - commands available after installation
# Useful for local development
[project.scripts]
stream-middleware = "stream.middleware.app:main"

[project.urls]
# Homepage = ""
Repository = "https://github.com/uicacer/STREAM"


[tool.ruff]
# Same as Black's line length
line-length = 100

# Target Python 3.11 (matches your requires-python)
target-version = "py311"

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "*.egg-info",
    "dist",
    "build",
    ".pytest_cache",
    ".ipynb_checkpoints",
]

[tool.ruff.lint]
# Enable these rule sets
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "UP",  # pyupgrade (modern Python syntax)
    "B",   # flake8-bugbear (common bugs)
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

# Ignore these specific rules
ignore = [
    "E501",  # line too long (formatter handles this)
    "B008",  # function calls in argument defaults (common in FastAPI)
    "N802",  # function name should be lowercase (allows mixed case)
]

# Allow unused variables when they start with underscore
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"stream/middleware/app.py" = ["E402"]  # Allow imports after code (spinner UX)

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with spaces (4 spaces)
indent-style = "space"

# Use Unix line endings
line-ending = "lf"
